requirements.txt
# FastAPI и зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Валидация и сериализация
pydantic==2.5.0
pydantic-settings==2.1.0

# Утилиты
python-slugify==8.0.1
python-dotenv==1.0.0

# CORS и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Для работы с датами
python-dateutil==2.8.2

# Логирование
structlog==23.2.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Разработка (опционально)
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Для админ панели (позже)
jinja2==3.1.2
aiofiles==23.2.1